+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: sksioui <sksioui@student.1337.ma>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/01 13:44:27 by sksioui           #+#    #+#             */
-/*   Updated: 2025/11/04 16:29:25 by sksioui          ###   ########.fr       */
+/*   Updated: 2025/11/11 10:47:22 by sksioui          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ int	ft_print_char(char c);
 int	ft_print_str(char *s);
 int	ft_print_nbr(int nb);
 int	ft_print_unsigned(unsigned int n);
-int	ft_print_hex(unsigned long n, char b);
+int	ft_print_hex(unsigned int n, char b);
 int	ft_print_ptr(void *ptr);
 
 #endif
